requirements.txt
assemblyai
fire
python-dotenv
mcp[cli]
uv 
nest_asyncio